Seed roles, permissions, audits, and operational types
Refactored repository interfaces to simplify generics. Migrated and updated the Audit entity and namespace. Changed Roles PK to short. Added EF Core configuration and seed data for roles, permissions, role-permissions, audit types, and operational change types. Introduced migration to create and seed related tables. Improved DbContext configuration and fixed logic in authorization and airport API key generation. Minor adjustments to error codes and logging structure.

+++ Core/AeroVeloz.Domain/Entities/Users/Roles/Roles.cs
@@ -7,7 +7,7 @@ namespace AeroVeloz.Domain.Entities.Users.Roles
     /// Los roles determinan el nivel de acceso y las acciones que un usuario puede realizar
     /// (ej: SYSTEMADMIN, AIRPORTADMIN, AIRLINEADMIN).
     /// </summary>
-    public partial class Roles : BEntity<int>
+    public partial class Roles : BEntity<short>
     {
         public string? nameRol { get; init; }
     }
